requirements.txt
google-adk[database]==0.3.0
yfinance==0.2.25
pandas==2.2.3
litellm==1.78.0
google-generativeai==0.8.5
python-dotenv==1.0.0
aiofiles>=0.8.0
aiohttp>=3.8.0
aiomqtt
docling
mcp-agent
mcp-server-git
nest_asyncio
pathlib2
PyPDF2>=2.0.0
reportlab>=3.5.0
streamlit
openai>=1.65.4
transformers>=4.46.3
tiktoken>=0.9.0
fastapi
uvicorn